+++ zarrs/src/array/codec/bytes_to_bytes/adler32.rs
@@ -75,7 +75,7 @@ mod tests {
             },
             BytesRepresentation,
         },
-        byte_range::ByteRange,
+        storage::byte_range::ByteRange,
     };
 
     use super::*;